+++ expects/matchers/built_in/__init__.py
@@ -12,6 +12,7 @@ from .be_below import be_below
 from .be_above_or_equal import be_above_or_equal
 from .be_below_or_equal import be_below_or_equal
 from .be_within import be_within
+from .be_callable import be_callable
 from .have_length import have_length
 from .have_properties import have_properties, have_property
 from .have_keys import have_key, have_keys
@@ -35,6 +36,7 @@ __all__ = [
     'be_above_or_equal',
     'be_below_or_equal',
     'be_within',
+    'be_callable',
     'have_length',
     'have_properties',
     'have_property',